--- src/main/java/nextstep/ladder/domain/exceptions/CannotNullOrEmptyException.java
@@ -1,9 +1,0 @@
-package nextstep.ladder.domain.exceptions;
-
-public class CannotNullOrEmptyException extends IllegalArgumentException {
-    public CannotNullOrEmptyException() {
-        super("null이거나 비어있을 수 없습니다.");
-    }
-
-}
-
