+++ src/main/java/nextstep/ladder/exceptions/CannotNullOrEmptyException.java
@@ -1,0 +1,9 @@
+package nextstep.ladder.exceptions;
+
+public class CannotNullOrEmptyException extends IllegalArgumentException {
+    public CannotNullOrEmptyException() {
+        super("null이거나 비어있을 수 없습니다.");
+    }
+
+}
+
