+++ src/main/java/nextstep/ladder/exceptions/ExceedMaxNameLengthException.java
@@ -1,0 +1,8 @@
+package nextstep.ladder.exceptions;
+
+public class ExceedMaxNameLengthException extends IllegalArgumentException {
+    public ExceedMaxNameLengthException() {
+        super("이름 최대 길이를 초과했습니다.");
+    }
+
+}
