+++ src/main/java/nextstep/ladder/view/OutputView.java
@@ -1,9 +1,9 @@
 package nextstep.ladder.view;
 
-import nextstep.ladder.domain.Lines;
-import nextstep.ladder.domain.Members;
-import nextstep.ladder.domain.RunResult;
-import nextstep.ladder.domain.RunResults;
+import nextstep.ladder.domain.line.Lines;
+import nextstep.ladder.domain.member.Members;
+import nextstep.ladder.domain.runresult.RunResult;
+import nextstep.ladder.domain.runresult.RunResults;
 
 public class OutputView {
     public static void outputLadder(Members members, Lines lines, RunResults runResults) {
