+++ src/test/java/nextstep/ladder/domain/MaxHeightOfLadderTest.java
@@ -1,7 +1,11 @@
 package nextstep.ladder.domain;
 
-import nextstep.ladder.domain.exceptions.NotAllowNegativeOrZero;
-import nextstep.ladder.domain.exceptions.NotNumberStringIsNotAllowException;
+import nextstep.ladder.exceptions.NotAllowNegativeOrZero;
+import nextstep.ladder.exceptions.NotNumberStringIsNotAllowException;
+import nextstep.ladder.domain.line.Line;
+import nextstep.ladder.domain.line.LineTest;
+import nextstep.ladder.domain.line.LinesTest;
+import nextstep.ladder.domain.member.MembersTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
